+++ docs/changelog.rst
@@ -54,6 +54,8 @@ Detailed list of changes
 
 - A new action :ac:`send_key` to simplify mapping key presses to other keys without needing :ac:`send_text`
 
+- Allow focusing previously active OS windows via :ac:`nth_os_window` (:pull:`7009`)
+
 - Wayland: Fix a regression in the previous release that broke copying to clipboard under wl-roots based compositors in some circumstances
   (:iss:`6890`)